fitness-agent/requirements.txt
# Cloud Functions framework
functions-framework==3.*

# Google APIs
google-auth==2.23.0
google-api-python-client==2.108.0

# AI Providers
anthropic>=0.34.0
openai>=1.0.0
google-generativeai>=0.3.0

# Email
sendgrid>=6.10.0

# Utilities
python-dotenv>=1.0.0


